requirements.txt
camelot-py[base]
opencv-python
PyPDF2<3.0